requirements.txt
ollama
yfinance
requests
psycopg2-binary
numpy
scipy
python-dotenv
python-telegram-bot

